+++ backend/api.py
@@ -4,6 +4,8 @@ from pathlib import Path
 from dotenv import load_dotenv
 from . import __version__ as version
 from .auth import setup_auth, validated_user
+from .authlocal import setup_authlocal
+from .models import UserProfile, TokenUser
 
 load_dotenv()
 app = FastAPI()
@@ -27,11 +29,10 @@ def apiVersion():
     )
 
 @app.get('/api/me')
-def apiMe(user: dict = Depends(validated_user)):
-    return dict(
-        user,
-        avatar = user['picture'],
-        roles=['customer'],
+def apiMe(user: dict = Depends(validated_user)) -> UserProfile:
+    # TODO: Either query ERP or have a rich jwt and take data from it
+    default = dict(
+        avatar = user.get('picture', None),
         nif = '12345678X',
         address = 'Rue del Percebe, 13',
         city = 'Salt',
@@ -40,7 +41,10 @@ def apiMe(user: dict = Depends(validated_user)):
         phone = '555444333',
         proxy_name = 'Matute Gonzalez, Frasco',
         proxy_nif = '987654321X',
+        roles = ['customer'],
     )
+    return UserProfile(**dict(default, **user))
 
 setup_auth(app)
+setup_authlocal(app)