+++ frontend/src/ProfilePage/index.jsx
@@ -4,6 +4,7 @@ import Container from '@mui/material/Container'
 import Chip from '@mui/material/Chip'
 import Avatar from '@mui/material/Avatar'
 import TextField from '@mui/material/TextField'
+import Typography from '@mui/material/Typography'
 import { useTranslation } from 'react-i18next'
 import { useAuth } from '../components/AuthProvider'
 import PageGuard from '../components/PageGuard'
@@ -68,6 +69,9 @@ export default function ProfilePage(params) {
   console.log('Guarding for', currentUser)
   return (
     <Container>
+      <Typography variant="h3" sx={{ mb: 3 }}>
+        {t('PROFILE.PROFILE_TITLE')}
+      </Typography>
       <Box
         style={{
           display: 'grid',
@@ -91,16 +95,16 @@ export default function ProfilePage(params) {
         </Box>
         {fields.map((field) => {
           return (
-            <>
+            <React.Fragment key={field.id}>
               <Box sx={{ textAlign: 'right' }}>
-                <label key={field.id + '_label'} for={field.id}>
+                <label htmlFor={field.id}>
                   <b>{field.label}:</b>
                 </label>
               </Box>
-              <Box key={field.id}>
+              <Box id={field.id}>
                 {field.view ? field.view(currentUser) : currentUser[field.id]}
               </Box>
-            </>
+            </React.Fragment>
           )
         })}
       </Box>