+++ CHANGES.md
@@ -2,4 +2,6 @@
 
 ## Unreleased
 
-
+- Upgrade Notes:
+    - Added `ERP_PROVISIONING_APIKEY` to .env to
+      provide a provisioning key for the ERP.